requirements.txt
# ADK Greeting Chat Application - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Web Framework
streamlit>=1.28.0

# Environment Variables Management
python-dotenv>=1.0.0

# Google AI and ADK Dependencies
google-generativeai>=0.3.0
google-adk

# Async Processing Support
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0

# Typing Extensions (for better type hints)
typing-extensions>=4.8.0

# Additional Utilities
pydantic>=2.0.0
requests>=2.31.0
deprecated
psycopg2-binary

# Note: Exact versions may vary based on your Python version and system
# If you encounter version conflicts, try installing without version constraints first:
# pip install streamlit python-dotenv google-generativeai google-adk